Rename team js file to index
