+++ src/views/Team/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import SideNav from '../../components/SideNav';
-import Team from '../../components/Team/Team';
+import Team from '../../components/Team';
 
 const TeamView = () => (
   <>
